refactor: Split drawSeries.ts into smaller files
Extracted decimation caching logic to decimationCache.ts and overlay drawing logic to drawOverlay.ts to improve maintainability of drawSeries.ts.
